+++ src/sdk/CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 
+- Use ConcurrentDictionary to track batches in the BatchClient class [jansenbe - Bert Jansen]
 - Fix BannerImageUrl to use sphost only since siteid, webid are given as parameter #439 [czullu - Christian Zuellig]
 - Alignment of PageHeaderTitleAlignment enum with what's used in PnP Framework [jansenbe - Bert Jansen]
 - Hubsite joining/unjoining documention update #416 [plamber - Patrick Lamber]